Touches up code in model.php

+++ model.php
@@ -241,7 +241,7 @@
 					else
 						$list.="<td><a href=\"users.php?action=activate&id=".$id."\">Activate</a></td>";
 					$list.="<td><a href=\"users.php?action=reset_passwd&name=".$name."&email=".urlencode($email)."\">Reset Password</a></td>
-					<td><a href=\"admin.php?action=delete&id=".(string)$id."\">Delete</a></td></tr>";
+					<td><a href=\"admin.php?action=delete&id=".(string)$id."\" onclick=\"return confirm_delete();\">Delete</a></td></tr>";
 
 		/* Every 10th user is the last of the group of 10 */
 				if($counter%10==0)
